requirements.txt
playwright
langchain
langchain-core
langchain-openai
pydantic
deepagents
tavily-python
python-dotenv
Pillow
